\name{CA_OMICS_ENV}
\alias{CA_OMICS_ENV}
\docType{data}
\title{
    caOmicsV Environment
}
\description{
    The caOmicsV Environment holds and protects all parameters and objects 
    used for caOmicsV plot.
}
\usage{ CA_OMICS_ENV }
\format{
    The format is:
    <environment: 0x26b3e00> 
}
\value{
    None
}
\examples{
    is.environment(CA_OMICS_ENV)
}
\keyword{datasets}
