\name{CA_OMICS_NA_STRING}
\alias{CA_OMICS_NA_STRING}
\docType{data}
\title{
    The Default NA String Used by caOmicsV Package
}
\description{
    The default NA string used by caOmicV package is NULL
}
\usage{CA_OMICS_NA_STRING}
\format{
    The format is:
    NULL
}
\value{
    None
}
\examples{
    CA_OMICS_NA_STRING
}
\keyword{datasets}
