\name{CA_OMICS_NAME}
\alias{CA_OMICS_NAME}
\docType{data}
\title{
    The Name of caOmicsV Environment
}
\description{
    "CA_OMICS_ENV" is used for caOmicsV environment name.
}
\usage{CA_OMICS_NAME}
\format{
    The format is:
    chr "CA_OMICS_ENV"
}
\value{
    None
}
\examples{
    caOmicsVEnvironment <- get(CA_OMICS_NAME, envir=globalenv())
}
\keyword{datasets}
