\name{caOmicsV-package}
\alias{caOmicsV-package}
\alias{caOmicsV}
\docType{package}
\title{
    caOmicsV bioMatrix and bioNetCircos Layout Plot
}
\description{
    This package provides methods to display genomic data with two layouts:

    bioMatrix layout: sample info and multiple genomic data are displayed as 
    matrix with rows for features(phenotypes and genes, and columns for each 
    sample. Omics data will be plotted with different layers when necessary. 

    bioNetCircos layout: sample info and multiple genomic data are displayed 
    on an biological network with node for features (e.g., genes) and on each 
    node, sample info and genomic data are plotted with circular layout.
}
\author{
    Henry Zhang
    Maintainer: Henry Zhang <hzhang@mail.nih.gov>
}
\keyword{ package }
