\name{getCaOmicsVColors}
\alias{getCaOmicsVColors}
\title{
    Get Default Colors Used by caOmicsV Plot
}
\description{
    Display default colors used by caOmicsV package
}
\usage{
    getCaOmicsVColors()
}
\value{
    A character vector of length 8 for R colors including of "red", "blue", 
    "black", "green", "cyan", "brown", "magenta", and "gold".
}
\author{
    Henry Zhang
}
\examples{
    defaultColors <- getCaOmicsVColors()
}
\keyword{ methods }

