\name{getCaOmicsVPlotTypes}
\alias{getCaOmicsVPlotTypes}
\title{
    Get Plot Types Supported by caOmicsV Package }
\description{
    Retrieve the plot types supported by current version of caOmicsV package
}
\usage{
    getCaOmicsVPlotTypes()
}
\value{
    Character vector including "polygon", "bar", "points", "heatmap", "line", 
    "category", "binary"
}
\author{
    Henry Zhang
}
\examples{
    plotType <- getCaOmicsVPlotTypes()
}
\keyword{ methods }
