\name{getDefaultNaStrings}
\alias{getDefaultNaStrings}
\title{
    Default NA String
}
\description{
    Get the default NS strings used by caOmicsV package
}
\usage{
    getDefaultNaStrings()
}
\value{
    Returns na strings used by caOmicsV package. Default is "NULL".
}
\author{
    Henry Zhang
}
\examples{
    naStr <- getDefaultNaStrings()
}
\keyword{ methods }
