\name{miRNADemoData}
\alias{miRNADemoData}
\docType{data}
\title{
    miRNASeq Demo Data Set
}
\description{
    A data frame with miRNA read counts for 26 genes in 60 samples for heatmap 
    plot demo on both caOmicsV bioNetCircos layout and bioMatrix layout after 
    transformed to log2 values.
}
\usage{data("miRNADemoData")}
\format{
    A data frame with 26 observations on the following 61 variables.
    \describe{
        \item{\code{miRNA_ID}}{a character vector}
        \item{\code{TCGA.BC.A216.Normal}}{a numeric vector}
        \item{\code{TCGA.BD.A2L6.Normal}}{a numeric vector}
        \item{\code{TCGA.BD.A3EP.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A113.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A114.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A118.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A119.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A11A.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A11B.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A11C.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A11D.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EB.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EC.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EG.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EH.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EI.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EJ.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A1EL.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A39V.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A39W.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A39X.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A39Z.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A3A1.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A3A2.Normal}}{a numeric vector}
        \item{\code{TCGA.DD.A3A3.Normal}}{a numeric vector}
        \item{\code{TCGA.EP.A12J.Normal}}{a numeric vector}
        \item{\code{TCGA.EP.A26S.Normal}}{a numeric vector}
        \item{\code{TCGA.ES.A2HT.Normal}}{a numeric vector}
        \item{\code{TCGA.FV.A23B.Normal}}{a numeric vector}
        \item{\code{TCGA.FV.A2QR.Normal}}{a numeric vector}
        \item{\code{TCGA.BC.A216.Tumor}}{a numeric vector}
        \item{\code{TCGA.BD.A2L6.Tumor}}{a numeric vector}
        \item{\code{TCGA.BD.A3EP.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A113.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A114.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A118.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A119.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A11A.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A11B.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A11C.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A11D.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EB.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EC.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EG.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EH.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EI.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EJ.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A1EL.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A39V.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A39W.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A39X.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A39Z.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A3A1.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A3A2.Tumor}}{a numeric vector}
        \item{\code{TCGA.DD.A3A3.Tumor}}{a numeric vector}
        \item{\code{TCGA.EP.A12J.Tumor}}{a numeric vector}
        \item{\code{TCGA.EP.A26S.Tumor}}{a numeric vector}
        \item{\code{TCGA.ES.A2HT.Tumor}}{a numeric vector}
        \item{\code{TCGA.FV.A23B.Tumor}}{a numeric vector}
        \item{\code{TCGA.FV.A2QR.Tumor}}{a numeric vector}
    }
}
\value{
    None 
}
\examples{
    data(miRNADemoData)
}
\keyword{datasets}
