\name{setBioNetCircosBasePlotPositions}
\alias{setBioNetCircosBasePlotPositions}
\title{
    Set up Default Plot Positions for a Node on caOmicsV bioNetCircor Layout
}
\description{
    This function calculates default x and y coordinates for points on a 
    circular line. Degrees a character string should be rotated on each point 
    are also calculated. This function is for internal used only.
}
\usage{
    setBioNetCircosBasePlotPositions(totalSamples=100, sampleWidth=100)
}
\arguments{
    \item{totalSamples}{
    non-negative numeric, total number of samples to be plotted
    }
    \item{sampleWidth}{
    non-negative numeric, width (how many points) of a sample on the circular 
    line
    }
}
\value{
    None
}
\author{
    Henry Zhang
}
\examples{
#  Internal use only
}
\keyword{ methods }
