\name{setCaOmicsVColors}
\alias{setCaOmicsVColors}
\title{
Set up Default Plot Colors for caOmicsV Plot
}
\description{
    Set or change default colors used for caOmicsV Plot. 
}
\usage{
    setCaOmicsVColors(colorList=NULL)
}
\arguments{
    \item{colorList}{
    character vector, list of colour names or R colour specification
    }
}
\value{
    None
}
\author{
    Henry Zhang
}
\examples{
    setCaOmicsVColors(colorList=c("red", "blue", "black", "green", "cyan", 
            "purple", "magenta", "grey"))
}
\keyword{ methods }
