\name{setDefaultNaStrings}
\alias{setDefaultNaStrings}
\title{
    Set up Default NA Strings for caOmicsV Package
}
\description{
    Set or change default NA strings used by caOmicsV Package.
}
\usage{
    setDefaultNaStrings(nullStrings)
}
\arguments{
    \item{nullStrings}{
    character vector of null strings
    }
}
\value{
    None
}
\author{
    Henry Zhang
}
\examples{
    setDefaultNaStrings("null")
}
\keyword{ methods }
