\name{showSupportedBioNetCircosPlotType}
\alias{showSupportedBioNetCircosPlotType}
\title{
    Display the Plot Types Supported by caOmicsV bioNetCircos Plot 
}
\description{
    Display plot types supported by current version of caOmicsV bioNetCircos 
    plot. A bioNetCircos layout must be initialized first.
}
\usage{
    showSupportedBioNetCircosPlotType()
}
\value{
    None
}
\author{
    Henry Zhang
}
\examples{
    showSupportedBioNetCircosPlotType()
}
\keyword{ methods }
